openblas64 0.3.21 (new formula)
OpenBLAS uses 32-bit integers by default to follow the reference BLAS.
This results in inefficient memory usage on 64-bit machines.

We can configure `openblas` to use 64-bit integers instead, but this is
likely to break many dependents that expect the same ABI as the
reference BLAS. To avoid this breakage, we can ship a 64-bit OpenBLAS as
`openblas64`. This is similar to what other distros (e.g. Debian,
Fedora) do for OpenBLAS.

The 64-bit version of OpenBLAS is what is used upstream by projects such
as Julia and NumPy.

We configure this with `SYMBOLSUFFIX` set to `64_` to allow loading both
versions of OpenBLAS in the same process. This is also the configuration
used in the projects mentioned above.